self review cleanup/refactors
