+++ src/views/DEX/components/HistoryContainer/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./HistoryContainer";
